+++ src/core/index.ts
@@ -2,5 +2,6 @@ export { Stack } from "./stack";
 export { Vec2 } from "./vec2";
 export { LRUCache } from "./LRU-cache";
 export { SmallIntCache } from "./small-int-cache";
+export { IndexArray } from "./index-array";
 export { SignedIndexArray } from "./signed-index-array";
 export { Map1, Map2, Map3 } from "./map";